Convert parser to iterator
This has the benefit of avoiding extra heap allocations that we were
making when building the vector of tokens in the old parser.

It is also clear that we are consuming the iterator, the old parser
would modify self and calling it a second time would yield nothing,
which might be unexpected.